Removed move camera from birds eye camera
