+++ crates/camera/src/birds_eye.rs
@@ -4,8 +4,6 @@ use std::f32::consts::PI;
 
 use bevy::ecs::schedule::ShouldRun;
 
-use super::MainCamera;
-
 pub struct BirdsEyeCameraPlugin;
 
 #[derive(SystemLabel)]
@@ -18,11 +16,14 @@ impl Plugin for BirdsEyeCameraPlugin {
                 .with_run_criteria(is_active)
                 .with_system(target_moved)
                 .with_system(settings_changed)
-                .with_system(move_camera)
                 .label(BirdsEyeCameraUpdate),
         );
     }
 }
+
+#[derive(Component, Reflect, Default)]
+#[reflect(Component)]
+pub struct BirdsEyeCamera;
 
 #[derive(Component, Reflect, Default)]
 #[reflect(Component)]
@@ -49,8 +50,8 @@ pub fn is_active(config: Res<BirdsEyeCameraConfig>) -> ShouldRun {
 fn setup(mut commands: Commands) {
     commands.insert_resource(BirdsEyeCameraConfig {
         radial_distance: 10.0,
-        polar_angle: 0.0,     // Up-Down
-        azimuthal_angle: 0.0, // Left-Right
+        polar_angle: std::f32::consts::FRAC_PI_6,
+        azimuthal_angle: 0.0,
         rotate_speed: PI / 5.0,
         zoom_speed: 5.0,
         active: true,
@@ -64,15 +65,14 @@ fn target_moved(
         (
             With<BirdsEyeCameraTarget>,
             Changed<Transform>,
-            Without<MainCamera>,
+            Without<BirdsEyeCamera>,
         ),
     >,
-    mut q: Query<&mut Transform, With<MainCamera>>,
+    mut q: Query<&mut Transform, With<BirdsEyeCamera>>,
 ) {
     let target = match target.get_single() {
         Ok(t) => t,
-        Err(QuerySingleError::NoEntities(msg)) => {
-            trace!(msg);
+        Err(QuerySingleError::NoEntities(_)) => {
             return;
         }
         Err(QuerySingleError::MultipleEntities(_)) => {
@@ -81,14 +81,18 @@ fn target_moved(
     };
 
     if let Ok(mut camera_transform) = q.get_single_mut() {
+        trace!(
+            "Updating camera look and move since target transform has changed. Config: {:?}",
+            config
+        );
         look_and_move_around(&mut camera_transform, target.translation, &config);
     }
 }
 
 fn settings_changed(
     config: Res<BirdsEyeCameraConfig>,
-    target: Query<&Transform, (With<BirdsEyeCameraTarget>, Without<MainCamera>)>,
-    mut q: Query<&mut Transform, With<MainCamera>>,
+    target: Query<&Transform, (With<BirdsEyeCameraTarget>, Without<BirdsEyeCamera>)>,
+    mut q: Query<&mut Transform, With<BirdsEyeCamera>>,
 ) {
     if config.is_changed() == false {
         return;
@@ -96,8 +100,7 @@ fn settings_changed(
 
     let target = match target.get_single() {
         Ok(t) => t,
-        Err(QuerySingleError::NoEntities(msg)) => {
-            trace!(msg);
+        Err(QuerySingleError::NoEntities(_)) => {
             return;
         }
         Err(QuerySingleError::MultipleEntities(_)) => {
@@ -106,6 +109,10 @@ fn settings_changed(
     };
 
     if let Ok(mut camera_transform) = q.get_single_mut() {
+        trace!(
+            "Updating camera look and move since config has changed. Config: {:?}",
+            config
+        );
         look_and_move_around(&mut camera_transform, target.translation, &config);
     }
 }
@@ -115,14 +122,14 @@ fn look_and_move_around(
     target: Vec3,
     config: &BirdsEyeCameraConfig,
 ) {
-    camera_transform.look_at(target, Vec3::Y);
-
     camera_transform.translation = spherical_to_cartesian(
         config.radial_distance,
         config.polar_angle,
         config.azimuthal_angle,
         target,
     );
+
+    camera_transform.look_at(target, Vec3::Y);
 }
 
 fn spherical_to_cartesian(radius: f32, polar: f32, azimuth: f32, center: Vec3) -> Vec3 {
@@ -133,47 +140,4 @@ fn spherical_to_cartesian(radius: f32, polar: f32, azimuth: f32, center: Vec3) -
         radius * polar.sin(),
         polar_cos * azimuth.sin(),
     ) + center
-}
-
-fn move_camera(
-    input: Res<Input<KeyCode>>,
-    time: Res<Time>,
-    mut config: ResMut<BirdsEyeCameraConfig>,
-) {
-    let mut delta = Vec3::ZERO;
-
-    if input.pressed(KeyCode::Right) {
-        delta.x = -config.rotate_speed * time.delta_seconds();
-    } else if input.pressed(KeyCode::Left) {
-        delta.x = config.rotate_speed * time.delta_seconds();
-    }
-
-    if input.pressed(KeyCode::Up) {
-        delta.y = config.rotate_speed * time.delta_seconds();
-    } else if input.pressed(KeyCode::Down) {
-        delta.y = -config.rotate_speed * time.delta_seconds();
-    }
-
-    if input.pressed(KeyCode::PageUp) {
-        delta.z = -config.zoom_speed * time.delta_seconds();
-    } else if input.pressed(KeyCode::PageDown) {
-        delta.z = config.zoom_speed * time.delta_seconds();
-    }
-
-    if delta == Vec3::ZERO {
-        return;
-    }
-
-    config.azimuthal_angle += delta.x;
-    config.polar_angle += delta.y;
-    config.radial_distance += delta.z;
-
-    use std::f32::consts;
-
-    // config.azimuthal_angle = config.azimuthal_angle.clamp(0.0, consts::TAU);
-    config.polar_angle = config.polar_angle.clamp(
-        0.0 + consts::FRAC_PI_8,
-        consts::FRAC_PI_2 - consts::FRAC_PI_8,
-    );
-    config.radial_distance = config.radial_distance.clamp(1.0, 30.0);
 }
