+++ crates/camera/src/lib.rs
@@ -6,10 +6,6 @@ use self::fly_by::FlyByCameraPlugin;
 pub mod birds_eye;
 pub mod fly_by;
 
-#[derive(Component, Reflect, Default)]
-#[reflect(Component)]
-pub struct MainCamera;
-
 pub struct CameraPlugin;
 
 impl Plugin for CameraPlugin {
